add CowLoading component on storybook
