+++ src/apps/explorer/components/TransanctionBatchGraph/index.tsx
@@ -11,7 +11,6 @@ import CytoscapeComponent from 'react-cytoscapejs'
 import styled, { useTheme } from 'styled-components'
 import BigNumber from 'bignumber.js'
 
-import CowLoadingSVG from 'assets/img/cowLoading.svg'
 import { GetTxBatchTradesResult as TxBatchData, Settlement as TxSettlement } from 'hooks/useTxBatchTrades'
 import { networkOptions } from 'components/NetworkSelector'
 import { Network } from 'types'
@@ -22,7 +21,8 @@ import { APP_NAME } from 'const'
 import { HEIGHT_HEADER_FOOTER, TOKEN_SYMBOL_UNKNOWN } from 'apps/explorer/const'
 import { STYLESHEET } from './styled'
 import { abbreviateString, FormatAmountPrecision, formattingAmountPrecision } from 'utils'
-import { StyledCowLoading, WrapperCenter } from 'apps/explorer/pages/styled'
+import { WrapperCenter } from 'apps/explorer/pages/styled'
+import CowLoading from 'components/common/CowLoading'
 
 Cytoscape.use(popper)
 const PROTOCOL_NAME = APP_NAME
@@ -215,7 +215,7 @@ function TransanctionBatchGraph({
   if (isLoading)
     return (
       <WrapperCenter>
-        <StyledCowLoading src={CowLoadingSVG} />
+        <CowLoading />
       </WrapperCenter>
     )
