+++ src/apps/explorer/pages/styled.tsx
@@ -1,9 +1,8 @@
-import styled, { keyframes } from 'styled-components'
+import styled from 'styled-components'
 import { media } from 'theme/styles/media'
 import { RowWithCopyButton } from 'components/common/RowWithCopyButton'
 import * as CSS from 'csstype'
 import { Link } from 'react-router-dom'
-import SVG from 'react-inlinesvg'
 
 export const Wrapper = styled.div`
   padding: 1.6rem;
@@ -33,64 +32,6 @@ export const WrapperCenter = styled.div`
   align-items: center;
   justify-content: center;
   height: 100%;
-`
-
-const CowBounce = keyframes`
-  0%,
-    100% {
-      transform: scale(0.95);
-    }
-    50% {
-      transform: scale(1);
-    }
-`
-
-const EyesOpacity = keyframes`
-  from {
-      opacity: 1;
-    }
-    30% {
-      opacity: 0.3;
-    }
-
-    to {
-      opacity: 1;
-    }
-`
-
-export const StyledCowLoading = styled(SVG)`
-  animation: ${CowBounce} 1.5s infinite ease-in-out;
-  animation-delay: -1s;
-  path:nth-child(1) {
-    fill: ${({ theme }): string => theme.white};
-    opacity: 0.4;
-  }
-  path:nth-child(2) {
-    fill: ${({ theme }): string => theme.bg1};
-    opacity: 1;
-  }
-  path:nth-child(3) {
-    fill: ${({ theme }): string => theme.orange};
-    animation: ${EyesOpacity} 1.5s ease-in-out infinite;
-    filter: blur(1px);
-  }
-
-  /* path:nth-child(1) {
-    fill: ${({ theme }): string => theme.white};
-    opacity: 0.4;
-  }
-
-  path:nth-child(2),
-  path:nth-child(3) {
-    fill: ${({ theme }): string => theme.bg1};
-    opacity: 1;
-  }
-  path:nth-child(4),
-  path:nth-child(5) {
-    fill: ${({ theme }): string => theme.orange};
-    animation: ${EyesOpacity} 1.5s ease-in-out infinite;
-    filter: blur(1px);
-  } */
 `
 
 export const StyledTabLoader = styled.span`
